+++ project/include/naive_sort.h
@@ -1,0 +1,11 @@
+#ifndef naive_sort_h_
+#define naive_sort_h_
+
+#include "utils.h"
+
+#include <stdlib.h>
+
+void merge_sort(Array* array, size_t left, size_t right);
+
+
+#endif  // naive_sort_h_
