--- project/naive_include/naive_sort.h
@@ -1,10 +1,0 @@
-#ifndef naive_sort_h_
-#define naive_sort_h_
-
-#include "../utils/utils.h"
-
-#include <stdlib.h>
-
-void merge(Array* array, size_t left, size_t middle, size_t right);
-
-#endif  // naive_sort_h_
