+++ project/src/parallel_sort.c
@@ -8,10 +8,26 @@
 #define __USE_MISC
 #include <sys/mman.h>
 
+static int naive_merge_sort(Array* array, size_t left, size_t right) {
+    if (left >= right) {
+        return -1;
+    }
+
+    size_t middle = (left + right) / 2;
+
+    naive_merge_sort(array, left, middle);
+    naive_merge_sort(array, middle + 1, right);
+
+    merge(array, left, middle, right);
+    return 0;
+}
+
 
 void par_merge_sort(Array* array, size_t left, size_t right, int* nproc) {
     if (*nproc >= get_nprocs()) {
-        merge_sort(array, left, right);
+        if (naive_merge_sort(array, left, right) == -1) {
+            return;
+        }
         return;
     }
